+++ tests/document/__init__.py
@@ -1,4 +1,6 @@
-# TODO EXPLICT IMPORTS
+import sys
+sys.path[0:0] = [""]
+import unittest
 
 from class_methods import *
 from delta import *
@@ -6,6 +8,7 @@ from dynamic import *
 from indexes import *
 from inheritance import *
 from instance import *
+from json_serialisation import *
 
 if __name__ == '__main__':
     unittest.main()