+++ docs/changelog.rst
@@ -4,7 +4,9 @@ Changelog
 
 Changes in 0.8
 ==============
-- Updated index creation now tied to Document class ((MongoEngine/mongoengine#102)
+- Added to_json and from_json to Document (MongoEngine/mongoengine#1)
+- Added to_json and from_json to QuerySet (MongoEngine/mongoengine#131)
+- Updated index creation now tied to Document class (MongoEngine/mongoengine#102)
 - Added none() to queryset (MongoEngine/mongoengine#127)
 - Updated SequenceFields to allow post processing of the calculated counter value (MongoEngine/mongoengine#141)
 - Added clean method to documents for pre validation data cleaning (MongoEngine/mongoengine#60)